Move serializable trait to more appropriate module
Also make `Field` implement `Serializable`
