+++ src/bits/mod.rs
@@ -1,4 +1,5 @@
 use crate::secret_sharing::BooleanShare;
+use std::io;
 use std::ops::{BitAnd, BitAndAssign, BitOr, BitOrAssign, BitXor, BitXorAssign, Index, Not};
 
 pub mod bit_array;
@@ -34,4 +35,24 @@ impl<T> BooleanOps for T where
         + Not<Output = Self>
         + Sized
 {
+}
+
+/// Trait for items that have fixed-byte length representation.
+pub trait Serializable: Sized {
+    /// Required number of bytes to store this message on disk/network
+    const SIZE_IN_BYTES: usize;
+
+    /// Serialize this message to a mutable slice. Implementations need to ensure `buf` has enough
+    /// capacity to store this message.
+    ///
+    /// ## Errors
+    /// Returns an error if `buf` does not have enough capacity to store at least `SIZE_IN_BYTES` more
+    /// data.
+    fn serialize(self, buf: &mut [u8]) -> io::Result<()>;
+
+    /// Deserialize message from a sequence of bytes.
+    ///
+    /// ## Errors
+    /// Returns an error if the provided buffer does not have enough bytes to read (EOF).
+    fn deserialize(buf: &[u8]) -> io::Result<Self>;
 }
