+++ src/helpers/messaging.rs
@@ -7,7 +7,7 @@
 //! enables MPC protocols to do.
 //!
 use crate::{
-    ff::{Field, Int},
+    ff::Field,
     helpers::{
         buffers::{ReceiveBuffer, SendBuffer, SendBufferConfig},
         network::ChannelId,
@@ -19,11 +19,11 @@ use crate::{
 };
 use futures::StreamExt;
 use std::fmt::{Debug, Formatter};
-use std::io;
 use std::time::Duration;
 use tinyvec::array_vec;
 use tracing::Instrument;
 
+use crate::bits::Serializable;
 use crate::helpers::network::{MessageEnvelope, Network};
 use crate::helpers::time::Timer;
 use crate::helpers::transport::Transport;
@@ -32,40 +32,8 @@ use futures_util::stream::FuturesUnordered;
 #[cfg(all(feature = "shuttle", test))]
 use shuttle::future as tokio;
 
-/// Trait for items that have fixed-byte length representation.
-pub trait Serializable: Sized {
-    /// Required number of bytes to store this message on disk/network
-    const SIZE_IN_BYTES: usize;
-
-    /// Serialize this message to a mutable slice. Implementations need to ensure `buf` has enough
-    /// capacity to store this message.
-    ///
-    /// ## Errors
-    /// Returns an error if `buf` does not have enough capacity to store at least `SIZE_IN_BYTES` more
-    /// data.
-    fn serialize(self, buf: &mut [u8]) -> io::Result<()>;
-
-    /// Deserialize message from a sequence of bytes.
-    ///
-    /// ## Errors
-    /// Returns an error if the provided buffer does not have enough bytes to read (EOF).
-    fn deserialize(buf: &[u8]) -> io::Result<Self>;
-}
-
 /// Trait for messages sent between helpers. Everything needs to be serializable and safe to send.
 pub trait Message: Debug + Send + Serializable + 'static {}
-
-impl<F: Field> Serializable for F {
-    const SIZE_IN_BYTES: usize = (F::Integer::BITS / 8) as usize;
-
-    fn serialize(self, buf: &mut [u8]) -> io::Result<()> {
-        <F as Field>::serialize(&self, buf)
-    }
-
-    fn deserialize(buf: &[u8]) -> io::Result<Self> {
-        <F as Field>::deserialize(buf)
-    }
-}
 
 /// Any field value can be send as a message
 impl<F: Field> Message for F {}
