+++ src/secret_sharing/xor.rs
@@ -1,3 +1,4 @@
+use crate::bits::Serializable;
 use std::io;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -7,8 +8,6 @@ pub struct XorReplicated {
 }
 
 impl XorReplicated {
-    pub const SIZE_IN_BYTES: usize = 2 * std::mem::size_of::<u64>();
-
     #[must_use]
     pub fn new(left: u64, right: u64) -> Self {
         Self { left, right }
@@ -23,13 +22,26 @@ impl XorReplicated {
     pub fn right(&self) -> u64 {
         self.right
     }
+}
 
-    /// Deserializes this instance from a slice of bytes
-    ///
-    /// ## Errors
-    /// if buffer does not have enough capacity to hold a valid value of this instance.
-    #[allow(clippy::missing_panics_doc)]
-    pub fn deserialize(buf: &[u8]) -> io::Result<Self> {
+impl Serializable for XorReplicated {
+    const SIZE_IN_BYTES: usize = 2 * std::mem::size_of::<u64>();
+
+    fn serialize(self, buf: &mut [u8]) -> io::Result<()> {
+        if buf.len() < Self::SIZE_IN_BYTES {
+            Err(io::Error::new(
+                io::ErrorKind::WriteZero,
+                "not enough buffer capacity",
+            ))
+        } else {
+            buf[..8].copy_from_slice(&self.left.to_le_bytes());
+            buf[8..16].copy_from_slice(&self.right.to_le_bytes());
+
+            Ok(())
+        }
+    }
+
+    fn deserialize(buf: &[u8]) -> io::Result<Self> {
         if buf.len() < Self::SIZE_IN_BYTES {
             Err(io::Error::new(
                 io::ErrorKind::UnexpectedEof,
@@ -43,24 +55,6 @@ impl XorReplicated {
                 left: u64::from_le_bytes(left_buf),
                 right: u64::from_le_bytes(right_buf),
             })
-        }
-    }
-
-    /// Serializes this instance into a mutable slice of bytes, writing from index 0.
-    ///
-    /// ## Errors
-    /// if buffer capacity is not sufficient to hold all the bytes.
-    pub fn serialize(&self, buf: &mut [u8]) -> io::Result<usize> {
-        if buf.len() < Self::SIZE_IN_BYTES {
-            Err(io::Error::new(
-                io::ErrorKind::WriteZero,
-                "not enough buffer capacity",
-            ))
-        } else {
-            buf[..8].copy_from_slice(&self.left.to_le_bytes());
-            buf[8..16].copy_from_slice(&self.right.to_le_bytes());
-
-            Ok(Self::SIZE_IN_BYTES)
         }
     }
 }
